Initial Commit
This is my work thus far. The code needs cleaning up and the OAuth stuff is all tied to my personal gmail account as it doesn't seem that I can make my school account play nicely with the Google API .
The CSS, HTML, and JS probably needs cleaning up to be completed.